+++ benchmarks/IfItQuacks.Benchmarks/Program.cs
@@ -1,0 +1,3 @@
+using BenchmarkDotNet.Running;
+
+BenchmarkSwitcher.FromAssembly(typeof(IfItQuacks.Benchmarks.CallBenchmarks).Assembly).Run(args);